Enhance class TimeDependentCorrelation
Enhance the setup of the operators in the TimeDependenCorrelation class. Allow a similar
dictionary configuration for ``operator_t`` as already for ``operator_t0``.
It is now possible to arbitrary product operators as ``operator_t0`` *and* ``operator_t`` acting on multiple sites,
previously ``operator_t`` was restricted to be an onsite operator. Furthermore the configuration of ``operator_t``
can now be dropped, in which case the hermitian conjugate of the ``operator_t0`` is used.
If both operators act on multiple sites, correlations are still calculated by sliding the ``operator_t`` through
the MPS winding (starting at site 0) and by moving one site to the left. It is not yet supported to only calculate
the overlap on every n-th site.
Include a 'correlation_function_key' to name the measurement result accordingly.
Furthermore add "mixed-space" operators as (experimental) feature (with a more rudimentary setup for the operator
configurations). Update tests accrodingly.
Allow the spatial fourier transform to take an axis in spectral_function_tools.py and include a mixed_space parameter.